+++ src/cli/env.js
@@ -1,5 +1,9 @@
 const parseEnv = () => {
-    // Write your code here 
+    const enviromentVariables = Object.entries(process.env)
+        .filter(([key]) => key.startsWith("RSS"))
+        .map(([key, value]) => `${key}=${value}`)
+        .join("; ");
+    console.log(enviromentVariables);
 };
 
 parseEnv();
