+++ src/fs/copy.js
@@ -1,6 +1,6 @@
 import fs from "node:fs/promises";
 import path from "node:path";
-import * as url from "url";
+import * as url from "node:url";
 
 const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
